fix missing natcapd-client or mosquitto_sub

+++ natcap/files/natcapd.sh
@@ -171,16 +171,18 @@ enabled="`uci get natcapd.default.enabled 2>/dev/null`"
 		timeout -t5 sh -c 'echo >/tmp/trigger_gfwlist_update.fifo'
 	fi
 
-	#reload natcapd-client
-	natcap_redirect_port=`uci get natcapd.default.natcap_redirect_port 2>/dev/null || echo 0`
-	sleep 1 && killall natcapd-client >/dev/null 2>&1 && sleep 2
-	echo natcap_redirect_port=$natcap_redirect_port >$DEV
-	test $natcap_redirect_port -gt 0 && test $natcap_redirect_port -lt 65535 && {
-		(
-		/usr/sbin/natcapd-client -l$natcap_redirect_port >/dev/null 2>&1
-		echo natcap_redirect_port=0 >/dev/natcap_ctl
-		) &
-	}
+	if which natcapd-client >/dev/null 2>&1; then
+		#reload natcapd-client
+		natcap_redirect_port=`uci get natcapd.default.natcap_redirect_port 2>/dev/null || echo 0`
+		sleep 1 && killall natcapd-client >/dev/null 2>&1 && sleep 2
+		echo natcap_redirect_port=$natcap_redirect_port >$DEV
+		test $natcap_redirect_port -gt 0 && test $natcap_redirect_port -lt 65535 && {
+			(
+			/usr/sbin/natcapd-client -l$natcap_redirect_port >/dev/null 2>&1
+			echo natcap_redirect_port=0 >$DEV
+			) &
+		}
+	fi
 }
 
 #reload pptpd
@@ -300,11 +302,15 @@ txrx_vals() {
 mqtt_cli() {
 	while :; do
 		test -f $LOCKDIR/$PID || exit 0
-		mosquitto_sub -h router-sh.ptpt52.com -t "/gfw/device/$CLI" -u ptpt52 -P 153153 --quiet -k 180 | while read _line; do
+		if which mosquitto_sub >/dev/null 2>&1; then
+			mosquitto_sub -h router-sh.ptpt52.com -t "/gfw/device/$CLI" -u ptpt52 -P 153153 --quiet -k 180 | while read _line; do
+				timeout -t5 sh -c 'echo >/tmp/trigger_natcapd_update.fifo'
+			done
+			sleep 60
 			timeout -t5 sh -c 'echo >/tmp/trigger_natcapd_update.fifo'
-		done
-		sleep 60
-		timeout -t5 sh -c 'echo >/tmp/trigger_natcapd_update.fifo'
+		else
+			sleep 60
+		fi
 	done
 }
